+++ apps/web/src/routes/workspace-settings.test.tsx
@@ -97,7 +97,7 @@ GlobalRegistrator.register();
 (globalThis as typeof globalThis & { IS_REACT_ACT_ENVIRONMENT: boolean }).IS_REACT_ACT_ENVIRONMENT =
   true;
 
-const { MembersSection } = await import("./workspace-settings");
+const { MembersSection } = await import("./workspace-members-section");
 
 afterAll(() => {
   mock.restore();
