+++ scripts/web-bundle-budget-policy.test.ts
@@ -10,6 +10,8 @@ import {
   VARIABLE_SET_SELECTION_MERGE_TREE_RAW_MEASUREMENT,
   WORK_DISCOVERY_MERGE_TREE_RAW_BUDGET,
   WORK_DISCOVERY_MERGE_TREE_RAW_MEASUREMENT,
+  WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET,
+  WORKSPACE_MEMBER_ADMINISTRATION_RAW_MEASUREMENT,
   wholeKibEnvelope,
 } from "./web-bundle-budget-policy";
 
@@ -52,11 +54,20 @@ describe("web bundle budget policy", () => {
     expect(WORK_DISCOVERY_MERGE_TREE_RAW_BUDGET - WORK_DISCOVERY_MERGE_TREE_RAW_MEASUREMENT).toBe(
       1_632,
     );
+  });
+
+  test("retains the exact workspace-member administration envelope", () => {
+    expect(WORKSPACE_MEMBER_ADMINISTRATION_RAW_MEASUREMENT).toBe(2_210_226);
+    expect(WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET).toBe(2160 * KIB);
+    expect(
+      WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET - WORKSPACE_MEMBER_ADMINISTRATION_RAW_MEASUREMENT,
+    ).toBe(1_614);
     expect(EFFECTIVE_DIRECT_SESSION_RAW_BUDGET).toBe(
       Math.max(
         DIRECT_SESSION_RAW_BUDGET,
         VARIABLE_SET_SELECTION_MERGE_TREE_RAW_BUDGET,
         WORK_DISCOVERY_MERGE_TREE_RAW_BUDGET,
+        WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET,
       ),
     );
   });
