+++ scripts/web-bundle-budget-policy.ts
@@ -29,8 +29,15 @@ export const WORK_DISCOVERY_MERGE_TREE_RAW_MEASUREMENT = 2_206_112;
 export const WORK_DISCOVERY_MERGE_TREE_RAW_BUDGET = wholeKibEnvelope(
   WORK_DISCOVERY_MERGE_TREE_RAW_MEASUREMENT,
 );
+
+/** Exact workspace-member administration Linux/x64 Bun 1.4 merge-tree measurement. */
+export const WORKSPACE_MEMBER_ADMINISTRATION_RAW_MEASUREMENT = 2_210_226;
+export const WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET = wholeKibEnvelope(
+  WORKSPACE_MEMBER_ADMINISTRATION_RAW_MEASUREMENT,
+);
 export const EFFECTIVE_DIRECT_SESSION_RAW_BUDGET = Math.max(
   DIRECT_SESSION_RAW_BUDGET,
   VARIABLE_SET_SELECTION_MERGE_TREE_RAW_BUDGET,
   WORK_DISCOVERY_MERGE_TREE_RAW_BUDGET,
+  WORKSPACE_MEMBER_ADMINISTRATION_RAW_BUDGET,
 );
